requirements.txt
    langchain
    python-dotenv  
    langchain-community
    langchain-text-splitters
    langchain-openai
    ipykernel
    bs4
    faiss-cpu
    
    streamlit
    langchain_groq
    sentence_transformers
    langchain-chroma
    langchain_huggingface
    sentence_transformers 
    duckdb
    pandas 
    openai
    pypdf

    pymupdf

    langchain_ollama
    huggingface_hub
    fastapi
    uvicorn

    python-multipart

